feat: selective NamespaceSchema field MST coverage
Move and use the NamespaceContentHash to restrict the field coverage of
merkle tree content hashes over NamespaceSchema's.

Only the schema elements are to be part of the anti-entropy protocol (at
first!) as only these fields are CRDTs / deterministically merge-able.
Therefore the MST hashes must only consider these fields when deciding
if two nodes are consistent, as only these fields will be converged.
